requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Bcrypt>=1.0.0
Werkzeug>=2.3.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0
python-dotenv>=1.0.0